Separate StorageDir
